Added more OOP
