+++ data/dude.txt
@@ -1,0 +1,2 @@
+T, false, read book
+E, true, hello /from 2021-02-22T01:00 /to 2023-12-12T23:59
