+++ src/main/java/Task.java
@@ -64,23 +64,3 @@ public class Task {
     }
 
 }
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
-
